(hel-431) refactor + add tests
