+++ .env.test
@@ -16,5 +16,5 @@ DNUM_SFTP_PRIVATE_KEY=test
 DNUM_SFTP_USERNAME=test
 
 # On active les features pour les tests
-NEXT_PUBLIC_ENABLED_FEATURES=COMPTE_RESULTAT_EJ, DECONNEXION, BOUTON_RECHERCHE
+NEXT_PUBLIC_ENABLED_FEATURES=COMPTE_RESULTAT_EJ, DECONNEXION, BOUTON_RECHERCHE, CARTO_FRANCE_METROPOLE
